Move clock config to clocks.rs
